// lib/schemas.ts
export const cardVisualSchema = z.object({
  frontPath: z.string(),
  title: z.string(),
  backPath: z.string(),
  width: z.number().default(1348),
  height: z.number().default(850),
  organizationId: z.string(),
  id: z.string(),
  createdAt: z.date(),
  updatedAt: z.date(),
  deletedAt: z.date().optional(),
});

export type CardVisual = z.infer<typeof cardVisualSchema>;